# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scripts
    description: AI-generated video scripts
  - name: user
    description: User profile and billing
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /scripts:
    get:
      operationId: listScripts
      tags: [scripts]
      summary: List user scripts
      description: Returns paginated list of user's generated scripts
      security:
        - bearerAuth: []
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [TikTok, Instagram, YouTube]
          description: Filter by platform
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of scripts
          content:
            application/json:
              schema:
                type: object
                properties:
                  scripts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Script"
                  total:
                    type: integer
                required:
                  - scripts
                  - total
    post:
      operationId: generateScript
      tags: [scripts]
      summary: Generate a new script with AI
      description: Uses GPT-4o-mini to generate a structured viral video script
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScriptBody"
      responses:
        "201":
          description: Generated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "402":
          description: Free limit reached — upgrade required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitReachedError"

  /scripts/stats:
    get:
      operationId: getScriptStats
      tags: [scripts]
      summary: Get script generation stats
      description: Returns totals and platform breakdown for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptStats"

  /scripts/{id}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get a specific script
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "404":
          description: Not found
    delete:
      operationId: deleteScript
      tags: [scripts]
      summary: Delete a script
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /user/checkout:
    post:
      operationId: createCheckout
      tags: [user]
      summary: Create a Stripe checkout session for Pro upgrade
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required:
                  - url

  /user/stripe-webhook:
    post:
      operationId: stripeWebhook
      tags: [user]
      summary: Stripe webhook handler
      responses:
        "200":
          description: Acknowledged

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Script:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        topic:
          type: string
        platform:
          type: string
          enum: [TikTok, Instagram, YouTube]
        title:
          type: string
        hook:
          type: string
        body:
          type: string
        callToAction:
          type: string
        script:
          type: string
        hashtags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - topic
        - platform
        - title
        - hook
        - body
        - callToAction
        - script
        - hashtags
        - createdAt

    GenerateScriptBody:
      type: object
      properties:
        topic:
          type: string
          description: The video topic
        platform:
          type: string
          enum: [TikTok, Instagram, YouTube]
          description: Target platform
      required:
        - topic
        - platform

    ScriptStats:
      type: object
      properties:
        total:
          type: integer
        byPlatform:
          type: object
          properties:
            TikTok:
              type: integer
            Instagram:
              type: integer
            YouTube:
              type: integer
          required:
            - TikTok
            - Instagram
            - YouTube
        thisWeek:
          type: integer
      required:
        - total
        - byPlatform
        - thisWeek

    UserProfile:
      type: object
      properties:
        id:
          type: string
        isPro:
          type: boolean
        scriptsRemaining:
          type: integer
        stripeCustomerId:
          type: string
          nullable: true
      required:
        - id
        - isPro
        - scriptsRemaining

    LimitReachedError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        scriptsRemaining:
          type: integer
      required:
        - error
        - code
        - scriptsRemaining