# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns paginated list of user's generated scripts
 * @summary List user scripts
 */
export const listScriptsQueryLimitDefault = 20;
export const listScriptsQueryOffsetDefault = 0;

export const ListScriptsQueryParams = zod.object({
  platform: zod
    .enum(["TikTok", "Instagram", "YouTube"])
    .optional()
    .describe("Filter by platform"),
  limit: zod.coerce.number().default(listScriptsQueryLimitDefault),
  offset: zod.coerce.number().default(listScriptsQueryOffsetDefault),
});

export const ListScriptsResponse = zod.object({
  scripts: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      topic: zod.string(),
      platform: zod.enum(["TikTok", "Instagram", "YouTube"]),
      title: zod.string(),
      hook: zod.string(),
      body: zod.string(),
      callToAction: zod.string(),
      script: zod.string(),
      hashtags: zod.array(zod.string()),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * Uses GPT-4o-mini to generate a structured viral video script
 * @summary Generate a new script with AI
 */
export const GenerateScriptBody = zod.object({
  topic: zod.string().describe("The video topic"),
  platform: zod
    .enum(["TikTok", "Instagram", "YouTube"])
    .describe("Target platform"),
});

/**
 * Returns totals and platform breakdown for the current user
 * @summary Get script generation stats
 */
export const GetScriptStatsResponse = zod.object({
  total: zod.number(),
  byPlatform: zod.object({
    TikTok: zod.number(),
    Instagram: zod.number(),
    YouTube: zod.number(),
  }),
  thisWeek: zod.number(),
});

/**
 * @summary Get a specific script
 */
export const GetScriptParams = zod.object({
  id: zod.coerce.number(),
});

export const GetScriptResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  topic: zod.string(),
  platform: zod.enum(["TikTok", "Instagram", "YouTube"]),
  title: zod.string(),
  hook: zod.string(),
  body: zod.string(),
  callToAction: zod.string(),
  script: zod.string(),
  hashtags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a script
 */
export const DeleteScriptParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get current user profile
 */
export const GetUserProfileResponse = zod.object({
  id: zod.string(),
  isPro: zod.boolean(),
  scriptsRemaining: zod.number(),
  stripeCustomerId: zod.string().nullish(),
});

/**
 * @summary Create a Stripe checkout session for Pro upgrade
 */
export const CreateCheckoutResponse = zod.object({
  url: zod.string(),
});